+++ python_modules/dagster/dagster_tests/core_tests/engine_tests/test_multiprocessing.py
@@ -178,10 +178,9 @@ def define_subdag_pipeline():
         done = False
         while not done:
             time.sleep(0.25)
-            with open(context.solid_config, 'r') as fd:
-                if fd.read() == '111':
-                    done = True
-                    time.sleep(1)
+            if os.path.isfile(context.solid_config):
+                done = True
+                time.sleep(1)
 
     @solid(
         input_defs=[InputDefinition('after', Nothing)],
@@ -207,17 +206,20 @@ def test_seperate_sub_dags():
     ).build_pipeline_definition()
 
     with seven.TemporaryDirectory() as tempdir:
-        filename = os.path.join(tempdir, 'foo')
+        file_one = os.path.join(tempdir, 'foo_one')
+        file_two = os.path.join(tempdir, 'foo_two')
+        file_three = os.path.join(tempdir, 'foo_three')
+
         result = execute_pipeline(
             pipe,
             environment_dict={
                 'storage': {'filesystem': {}},
                 'execution': {'multiprocess': {'config': {'max_concurrent': 4}}},
                 'solids': {
-                    'waiter': {'config': filename},
-                    'counter_1': {'config': filename},
-                    'counter_2': {'config': filename},
-                    'counter_3': {'config': filename},
+                    'waiter': {'config': file_three},
+                    'counter_1': {'config': file_one},
+                    'counter_2': {'config': file_two},
+                    'counter_3': {'config': file_three},
                 },
             },
             instance=DagsterInstance.local_temp(),